解决冲突  收藏功能和改的订单参数

+++ app/src/main/java/zzbcar/cckj/com/nzzb/bean/OrderBean.java
@@ -100,7 +100,7 @@ public class OrderBean implements Serializable {
         private String takeMapLocation;
         private Object takeTime;
         private double totalAmount;
-        private int trafficDepositMoney;
+        private double trafficDepositMoney;
         private int type;
         private int userId;
 
@@ -296,7 +296,7 @@ public class OrderBean implements Serializable {
             this.totalAmount = totalAmount;
         }
 
-        public int getTrafficDepositMoney() {
+        public double getTrafficDepositMoney() {
             return trafficDepositMoney;
         }
